refactor: convert TimeEntries into ts
